+++ HISTORY.rst
@@ -20,6 +20,10 @@ New features and enhancements
 * New argument `var_as_string` for `get_cat_attrs` to return variable names as strings. (:pull:`233`).
 * New argument `copy` for `move_and_delete`. (:pull:`233`).
 * New argument `restrict_year` for `compute_indicators`. (:pull:`233`).
+* Add more comments in the template. (:pull:`233`, :issue:`232`).
+* ``generate_weights`` now allows to split weights between experiments, and make them vary along the time/horizon axis. (:issue:`108`, :pull:`231`).
+* New independence_level, `institution`, added to ``generate_weights``. (:pull:`231`).
+* Updated ``produce_horizon`` so it can accept multiple periods or warming levels. (:pull:`231`).
 * Add more comments in the template. (:pull:`233`,:pull:`235`, :issue:`232`).
 
 Breaking changes
@@ -29,6 +33,11 @@ Breaking changes
 * ``driving_institution`` was removed from the "default" xscen columns. (:pull:`222`).
 * Folder parsing utilities (``parse_directory``) moved to ``xscen.catutils``. Signature changed : ``globpattern`` removed, ``dirglob`` added, new ``patterns`` specifications. See doc for all changes. (:pull:`205`).
 * ``compute_indicators`` now returns all outputs produced by indicators with multiple outputs (such as `rain_season`). (:pull:`228`).
+* In ``generate_weights``, independence_level `all` was renamed `model`. (:pull:`231`).
+* In response to a bugfix, results for ``generate_weights(independence_level='GCM')`` are significantly altered. (:issue:`230`, :pull:`231`).
+* Legacy support for `stats_kwargs` in ``ensemble_stats`` was dropped. (:pull:`231`).
+* `period` in ``produce_horizon`` has been deprecated and replaced with `periods`. (:pull:`231`).
+* Some automated `to_level` were updated to reflect more recent changes. (:pull:`231`).
 
 Bug fixes
 ^^^^^^^^^
@@ -36,6 +45,7 @@ Bug fixes
 * Added NotImplemented errors when trying to call `climatological_mean` and `compute_deltas` with daily data. (:pull:`187`).
 * Minor documentation fixes. (:issue:`223`, :pull:`225`).
 * Fixed a bug in ``unstack_dates`` where it failed for anything other than seasons. (:pull:`228`).
+* ``cleanup`` with `common_attrs_only` now works even when no `cat` attribute is present in the datasets. (:pull:`231`).
 
 Internal changes
 ^^^^^^^^^^^^^^^^